requirements.txt

fastapi
uvicorn
python-dotenv
google-generativeai
pypdf
requests
multipart
streamlit
